% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixtureNormalD.R
\name{MixtureNormalDistribution}
\alias{MixtureNormalDistribution}
\title{Generate n observations from a mixture of two normal distributions}
\usage{
MixtureNormalDistribution(p1, mu1, mu2, sigma1, sigma2)
}
\arguments{
\item{p1}{probability}
}
\description{
This function allows you to Generate n observations from a mixture of two normal distributions
}
\examples{
MixtureNormalDistribution(0.1, 5, 4, 1, 2)
}
\keyword{Bernoulli}
