% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVAForecast.R
\name{RVAForecast}
\alias{RVAForecast}
\title{A 10 Day RVA weather forecaster}
\usage{
RVAForecast(Initial)
}
\arguments{
\item{Initial}{What is the intial weather. E.G."Sunny"}
}
\description{
This function allows you to predict the weather ten days into the future
}
\examples{
RVAForecast("Sunny")
}
\keyword{Weather}
